+++ examples/list_post_votes.py
@@ -1,5 +1,6 @@
 
-## python examples/site.py db0
+## Get the latest post by user and list its likes. Required admin access
+## python examples/post_likes.py db0
 
 import os
 import argparse
@@ -33,8 +34,24 @@ if not lemmy_password:
 lemmy = Lemmy(f"https://{lemmy_domain}")
 if lemmy_username and lemmy_password:
     login = lemmy.log_in(lemmy_username, lemmy_password)
-site = lemmy.site.get()
-if site:
-    print(json.dumps(site, indent=4))
+user = lemmy.get_user(username=args.username, return_user_object = True)    
+if user:
+    uposts = user.get_latest_posts()
+    if not len(uposts):
+        print("User hasn't made any posts")
+    votes = []
+    scores = []
+    more_votes = []
+    page = 1 
+    # TODO: Make this into a built-in method once a post is a class
+    while len(more_votes) >= 50 or page == 1:
+        more_votes = lemmy.post.list_votes(uposts[0]["post"]["id"], page=page).get("post_likes", [])
+        page += 1
+        votes += more_votes
+        for v in more_votes:
+            scores.append(v["score"])
+    #print(json.dumps(votes, indent=4))
+    print(scores)
+    print(f"{len(votes)} Total votes counted")
 else:
-    print("Retrieval of site information failed")
+    print("no matching username found")