+++ examples/regapps_class.py
@@ -1,14 +1,10 @@
-## python examples/user.py db0
+## python examples/regapps_class
 
 import argparse
-import json
 import os
-
 from pythorhead import Lemmy
 
 arg_parser = argparse.ArgumentParser()
-arg_parser.add_argument("username", action="store")
-arg_parser.add_argument("content", action="store")
 arg_parser.add_argument(
     "-d",
     "--lemmy_domain",
@@ -50,12 +46,11 @@ lemmy_password = args.lemmy_password
 if not lemmy_password:
     lemmy_password = os.getenv("LEMMY_PASSWORD")
 
-lemmy = Lemmy(f"https://{lemmy_domain}")
+lemmy = Lemmy(f"https://{lemmy_domain}", raise_exceptions=True, request_timeout=2)
 if lemmy_username and lemmy_password:
-    lemmy.log_in(lemmy_username, lemmy_password)
-user = lemmy.user.get(username=args.username)
-if not user:
-    raise Exception("No valid username found")
-pm = lemmy.private_message(args.content, user["person_view"]["person"]["id"])
-if not pm:
-    print("Sending private message failed")
+    login = lemmy.log_in(lemmy_username, lemmy_password)
+regapps = lemmy.get_registration_applications(limit=10)
+if not regapps:
+    print("No registration applications found")
+for regapp in regapps:
+    print(f"Application from user '{regapp.creator.name}' {regapp.get_application_status()}")