+++ examples/user_class.py
@@ -1,14 +1,11 @@
 ## python examples/user.py db0
 
 import argparse
-import json
 import os
-
 from pythorhead import Lemmy
 
 arg_parser = argparse.ArgumentParser()
 arg_parser.add_argument("username", action="store")
-arg_parser.add_argument("content", action="store")
 arg_parser.add_argument(
     "-d",
     "--lemmy_domain",
@@ -50,12 +47,11 @@ lemmy_password = args.lemmy_password
 if not lemmy_password:
     lemmy_password = os.getenv("LEMMY_PASSWORD")
 
-lemmy = Lemmy(f"https://{lemmy_domain}")
+lemmy = Lemmy(f"https://{lemmy_domain}", raise_exceptions=True, request_timeout=2)
 if lemmy_username and lemmy_password:
-    lemmy.log_in(lemmy_username, lemmy_password)
-user = lemmy.user.get(username=args.username)
-if not user:
-    raise Exception("No valid username found")
-pm = lemmy.private_message(args.content, user["person_view"]["person"]["id"])
-if not pm:
-    print("Sending private message failed")
+    login = lemmy.log_in(lemmy_username, lemmy_password)
+user = lemmy.get_user(username=args.username)
+if user:
+    print(user.asjson(indent=4))
+else:
+    print("no matching username found")