+++ pythorhead/types/listing.py
@@ -6,3 +6,11 @@ class ListingType(str, Enum):
     Community = "Community"
     Local = "Local"
     Subscribed = "Subscribed"
+    ModeratorView = "ModeratorView"
+
+class PostListingType(str, Enum):
+    All = "All"
+    Community = "Community"
+    Local = "Local"
+    Subscribed = "Subscribed"
+    ModeratorView = "ModeratorView"